+++ tests/commands/test_compile.py
@@ -86,9 +86,7 @@ def test_command_version_mismatch(monkeypatch, caplog):
     assert caplog.records[0].levelname == 'CRITICAL'
     assert caplog.records[0].message == "item 1: version error: this package uses version 1.0, but earlier packages " \
                                         "used version 1.1\nTry upgrading packages to the same version:\n  cat file " \
-                                        "[file ...] | ocdskit upgrade 1.0:1.1 | ocdskit compile --package " \
-                                        "--versioned\nor set --schema to the URL or path of the release schema to " \
-                                        "use:\n  ocdskit compile --schema SCHEMA --package --versioned"
+                                        "[file ...] | ocdskit upgrade 1.0:1.1 | ocdskit compile --package --versioned"
     assert excinfo.value.code == 1